Add msg merkleblock

+++ wire/msg_block.py
@@ -14,7 +14,7 @@ MaxBlockPayload = 4000000
 
 # maxTxPerBlock is the maximum number of transactions that could
 # possibly fit into a block.
-maxTxPerBlock = (MaxBlockPayload / minTxPayload) + 1
+maxTxPerBlock = int((MaxBlockPayload / minTxPayload) + 1)
 
 # MaxBlockPayload is the maximum bytes a block message can be in bytes.
 # After Segregated Witness, the max block payload has been raised to 4MB.
